refactor: changed folders name and divided mainVision system in multiple folders
